(feat)ScrumGame
Created the enviroment downloaded images. Modified the TestClass.java (=Main). Modified the Fields.java(new functions, every field can drow itself and has its deafult attributs like the location and is it a wall or not. they also have the default setters getters) Map.java(functions to draw the map and get the field array for the child classes) Acotr.java(At the moment triing to figure out how do not get trought the walls and down from the map as a global solution)