feat(ScrumGame)
GameOver.png(Made a picture for the case when the hero dies) StageComplated.png(Made a picture when the stage is complated) *-key.png(edited the basic hero moves to show that he has a key) Built in the functions to follow the stages. Fixed some bugs. Prepared for the gameover case. Did the new stage basic things: levelups restore hps reposition monsters)